requirements.txt
certifi==2018.8.24
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
diff-match-patch==20121119
Django==2.0
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
django-crispy-forms==1.7.2
django-filter==2.0.0
django-formtools==2.1
django-guardian==1.4.9
django-haystack==2.8.1
django-import-export==1.0.1
django-mdeditor==0.1.8
django-redis==4.10.0
django-reversion==3.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
drf-extensions==0.4.0
drf-haystack==1.8.4
elasticsearch==2.4.1
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.7
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
Markdown==3.0.1
MarkupSafe==1.0
mysqlclient==1.3.14
odfpy==1.3.6
openpyxl==2.5.7
Pillow==5.2.0
psycopg2==2.7.6.1
PyJWT==1.6.4
python-dateutil==2.7.5
pytz==2018.5
PyYAML==3.13
redis==3.0.1
requests==2.19.1
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
xlrd==1.1.0
xlwt==1.3.0
